pyproject.toml
[project]
name = "project-r"
version = "0.1.0"
description = ""
authors = [
    {name = "puneeth-chowdary",email = "puneethvipparla@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "transformers (>=4.56.1,<5.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "qdrant-client (>=1.15.0)",
    "bson (>=0.5.10)"
]

[tool.poetry]
packages = [{include = "project_r", from = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
zenml = "0.84.3"
pymongo = "4.14.1"
click = "^8.0.1"
loguru = "^0.7.2"
rich = "^13.7.1"
numpy = "^1.26.4"
poethepoet = "0.29.0"
datasets = "^3.0.1"
torch = "2.2.2"

# Digital data ETL
selenium = "^4.21.0"
webdriver-manager = "^4.0.1"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
jmespath = "^1.0.1"
chromedriver-autoinstaller = "^0.6.4"

# Feature engineering
qdrant-client = "^1.8.0"
langchain = "0.2.12"
sentence-transformers = "5.1.0"

# RAG
langchain-openai = "^0.1.3"
jinja2 = "^3.1.4"
tiktoken = "^0.7.0"
fake-useragent = "^1.5.1"
langchain-community = "0.2.11"

# Inference
fastapi = "0.116.1"
uvicorn = "^0.30.6"
tenacity = ">=9.1.2,<10.0.0"
opik = "^0.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
pre-commit = "^3.7.1"
pytest = "^8.2.2"


[tool.poetry.group.aws.dependencies]
sagemaker = ">=2.232.2"
s3fs = ">2022.3.0"
aws-profile-manager = "^0.7.3"
kubernetes = "^30.1.0"
sagemaker-huggingface-inference-toolkit = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
